# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Frasco API",
    "version": "1.0.0",
    "description": "OpenAPI spec generated for Express Auth routes, to be used with ZAP DAST."
  },
  "servers": [
    {
      "url": "https://api.frascoengineer.com"
    }
  ],
  "paths": {
...
    "/webauthn/devices/{id}": {
      "delete": {
        "summary": "Delete a WebAuthn device",
        "security": [
          {
            "accessTokenCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device deleted"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "patch": {
        "summary": "Update label of a WebAuthn device",
        "security": [
          {
            "accessTokenCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string"
                  }
                },
                "required": [
                  "label"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Label updated"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "accessTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "JWT token stored in cookie after login"
      }
    }
  }
}